requirements.txt
asgiref==3.4.1
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.46.0
django-countries==7.2.1
django-debug-toolbar==3.2.2
django-etc==1.3.0
django-heroku==0.3.1
django-phonenumber-field==6.0.0
django-social-share==2.2.1
django-summernote==0.8.20.0
django-taggit==2.0.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
oauthlib==3.1.1
packaging==21.2
phonenumbers==8.12.37
Pillow==8.4.0
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
webencodings==0.5.1
whitenoise==6.0.0